+++ x/pocketcore/keeper/service.go
@@ -9,9 +9,8 @@ import (
 
 // this is the main call for a service node handling a relay request
 func (k Keeper) HandleRelay(ctx sdk.Ctx, relay pc.Relay) (*pc.RelayResponse, sdk.Error) {
-	ctx.Logger().Info(fmt.Sprintf("HandleRelay(Relay = %+v)", relay))
 	// get the latest session block height because this relay will correspond with the latest session
-	sessionBlockHeight := k.GetLatestSessionBlock(ctx).BlockHeight()
+	sessionBlockHeight := k.GetLatestSessionBlockHeight(ctx)
 	// retrieve all service nodes available from world state to do session generation (the session data is needed to service)
 	allNodes := k.GetAllNodes(ctx)
 	// get self node (your validator) from the current state
@@ -26,13 +25,18 @@ func (k Keeper) HandleRelay(ctx sdk.Ctx, relay pc.Relay) (*pc.RelayResponse, sdk
 	if !found {
 		return nil, pc.NewAppNotFoundError(pc.ModuleName)
 	}
+	// get the session context
+	sessionCtx, er := ctx.PrevCtx(sessionBlockHeight)
+	if er != nil {
+		return nil, sdk.ErrInternal(er.Error())
+	}
 	// ensure the validity of the relay
-	if err := relay.Validate(ctx, selfNode, hostedBlockchains, sessionBlockHeight, int(k.SessionNodeCount(ctx)), allNodes, app); err != nil {
-		ctx.Logger().Error(fmt.Errorf("could not validate for %v, %v, %v %v, %v, %v \n", selfNode, hostedBlockchains, sessionBlockHeight, int(k.SessionNodeCount(ctx)), allNodes, app).Error())
+	if err := relay.Validate(ctx, selfNode, hostedBlockchains, sessionBlockHeight, int(k.SessionNodeCount(sessionCtx)), allNodes, app); err != nil {
+		ctx.Logger().Error(fmt.Errorf("could not validate for %v, %v, %v %v, %v, %v \n", selfNode, hostedBlockchains, sessionBlockHeight, int(k.SessionNodeCount(sessionCtx)), allNodes, app).Error())
 		return nil, err
 	}
 	// store the proof before execution, because the proof corresponds to the previous relay
-	if err := relay.HandleProof(ctx, sessionBlockHeight); err != nil {
+	if err := relay.Proof.Handle(); err != nil {
 		ctx.Logger().Error(fmt.Errorf("could not handle proof for %v", sessionBlockHeight).Error())
 		return nil, err
 	}
@@ -59,4 +63,40 @@ func (k Keeper) HandleRelay(ctx sdk.Ctx, relay pc.Relay) (*pc.RelayResponse, sdk
 	}
 	resp.Signature = hex.EncodeToString(sig)
 	return resp, nil
+}
+
+func (k Keeper) HandleChallenge(ctx sdk.Ctx, challenge pc.ChallengeProofInvalidData) (*pc.ChallengeResponse, sdk.Error) {
+	// retrieve all service nodes available from world state to do session generation (the session data is needed to service)
+	allNodes := k.GetAllNodes(ctx)
+	// get self node (your validator) from the current state
+	selfNode, err := k.GetSelfNode(ctx)
+	if err != nil {
+		return nil, err
+	}
+	sessionBlkHeight := k.GetLatestSessionBlockHeight(ctx)
+	// get the session context
+	sessionCtx, er := ctx.PrevCtx(sessionBlkHeight)
+	if er != nil {
+		return nil, sdk.ErrInternal(er.Error())
+	}
+	// get the application that staked on behalf of the client
+	app, found := k.GetAppFromPublicKey(ctx, challenge.MinorityResponse.Proof.Token.ApplicationPublicKey)
+	// generate the session // todo caching
+	session, err := pc.NewSession(app.GetPublicKey().RawString(), challenge.MinorityResponse.Proof.Blockchain, pc.BlockHash(sessionCtx), sessionBlkHeight, allNodes, int(k.SessionNodeCount(sessionCtx)))
+	if err != nil {
+		return nil, err
+	}
+	if !found {
+		return nil, pc.NewAppNotFoundError(pc.ModuleName)
+	}
+	err = challenge.ValidateLocal(app.GetMaxRelays().Int64(), sessionBlkHeight, app.GetChains(), int(k.SessionNodeCount(sessionCtx)), session.SessionNodes, selfNode.GetAddress())
+	if err != nil {
+		return nil, err
+	}
+	// store the challenge in memory
+	err = challenge.Handle()
+	if err != nil {
+		return nil, err
+	}
+	return nil, nil
 }