+++ x/pocketcore/keeper/session.go
@@ -6,28 +6,29 @@ import (
 	sdk "github.com/pokt-network/posmint/types"
 )
 
-func (k Keeper) Dispatch(ctx sdk.Ctx, header types.SessionHeader) (*types.Session, sdk.Error) {
+func (k Keeper) Dispatch(ctx sdk.Ctx, header types.SessionHeader) (*types.DispatchResponse, sdk.Error) {
 	latestSessionBlockHeight := k.GetLatestSessionBlockHeight(ctx)
 	header.SessionBlockHeight = latestSessionBlockHeight
 	err := header.ValidateHeader()
 	if err != nil {
 		return nil, err
 	}
-	sessionCtx := ctx.MustGetPrevCtx(header.SessionBlockHeight)
+	sessionCtx, er := ctx.PrevCtx(header.SessionBlockHeight)
+	if er != nil {
+		return nil, sdk.ErrInternal(er.Error())
+	}
 	sessionBlkHeader := sessionCtx.BlockHeader()
 	nodes := k.GetAllNodes(ctx)
-	return types.NewSession(header.ApplicationPubKey, header.Chain, hex.EncodeToString(sessionBlkHeader.LastBlockId.Hash),
-		header.SessionBlockHeight, nodes, int(k.SessionNodeCount(ctx)))
+	session, err := types.NewSession(header.ApplicationPubKey, header.Chain, hex.EncodeToString(sessionBlkHeader.LastBlockId.Hash), header.SessionBlockHeight, nodes, int(k.SessionNodeCount(ctx)))
+	if err != nil {
+		return nil, err
+	}
+	return &types.DispatchResponse{Session: *session, BlockHeight: ctx.BlockHeight()}, nil
 }
 
 // is the context block a session block?
 func (k Keeper) IsSessionBlock(ctx sdk.Ctx) bool {
 	return ctx.BlockHeight()%k.posKeeper.SessionBlockFrequency(ctx) == 1
-}
-
-// get the most recent session block from the cont
-func (k Keeper) GetLatestSessionBlock(ctx sdk.Ctx) sdk.Context {
-	return ctx.MustGetPrevCtx(k.GetLatestSessionBlockHeight(ctx))
 }
 
 // get the most recent session block from the cont